ocr_engine/requirements.txt
# OCR Engines
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0
easyocr>=1.7.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
surya-ocr>=0.4.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.21.0
numpy>=1.24.0

# Layout Analysis
layoutparser>=0.3.4
# detectron2 - requires separate installation

# Language Detection
langdetect>=1.0.9
fasttext-wheel>=0.9.2

# Document Processing
pdf2image>=1.16.3
pymupdf>=1.23.0

# Machine Learning & Persistence
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.10.0
pandas>=2.0.0

# Database
# sqlite3 is built-in to Python

# Utilities
python-dateutil>=2.8.2